First Unix Upload
This is a basic configuration of the Whois program written in perl to
accept a list of data. However, this upload only is able to handle *nix
machines that have the whois application installed as there are system
calls in the program. A to do would be do have a pure perl
implementation of whois within this script to expand the whois
functionality and to have the script run on multiple architectures.